feat: add support for SWA learning rate scheduling and ensemble evaluation
- Introduce `swa_lr_start` and `swa_lr_final` for SWA learning rate scheduling
- Add ensemble evaluation and prediction functions
- Update train functions to handle multiple models and return detailed fold results
- Enhance configuration with new parameters for SWA and ensemble settings
- Refactor and optimize model training and evaluation logic

+++ csiro/metrics.py
@@ -26,7 +26,7 @@ def eval_global_wr2(model, dl_va, w_vec: torch.Tensor, device: str | torch.devic
             y_log = y_log.to(device, non_blocking=True)
             p_log = model(x).float()
 
-            y = torch.expm1(y_log)
+            y = torch.expm1(y_log.float())
             p = torch.expm1(p_log).clamp_min(0.0)
 
             w = w5.expand_as(y)
